pyproject.toml
[tool.poetry]
name = "pyreview"
version = "0.1.0"
description = ""
authors = ["od-barker <oliver.barker@housekeep.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.60.0"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.60.0"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
